requirements.txt
fastapi
fastapi-utils
uvicorn[standard]
typing-extensions
pydantic==1.10
python-dotenv
tensorflow
numpy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy[asyncio]
pymongo
weaviate-client
motor
redis
aioredis
celery
